BUGFIX: Inversion with artificial receivers runs!
Note that we do not see any sources emanating from
the artificial receiver locations, so we are in business.

TODO: (1) grad(div(p)), p=adjoint field
      (2) Integral of g(t) M grad(div(p))[s,t] over time
          for each receiver location -- that's the gradient in total

Above, g(t) refers to time signature, M=moment tensor, and
s is source location -- note that this will give us a 2d vector
and that is what we will update with!