BUGFIX for finite differencing, possibly still bugs
When I take a difference too small, I get divergence
of the differencing answer, but this might be natural
inaccuracy of the difference scheme I'm using.

Something to think about is (a) interpolating onto some
basis and then using known formulas or (b) asking around to
see what the state of the art is as far as this, (c) cubic
splines.