Add source perturbation -- BUGS in W2 adjoint!
Make perturbation in source function. We notice bug
in W2 adjoint for 50% perturbation (far from optimum).
W2 is registering 0! This should be the case where we are better smoothed.
Seems like 50% is significant amplitude loss case, so this may have to do with
it. This bug is CRITICAL and must be addressed before moving forward, or
we must understand the cases where this is going to happen.

TODO:
    (1) See if this can be remedied with different cutting parameter.
        What could be happening is that I am simply cutting too much information out.