Get rid of pre-windowing
NOTE: there is something fundamentally flawed with
pre-windowing when we construct `evaluators`.
If we only consider the support of g, we are
ignoring the support of f, which kind of negates
the idea of using optimal transport.

HOWEVER: note that our previous plots "sort of" worked
in that they captured the landscape for the "most part"
so we should not throw this windowing idea out entirely.
Furthermore, it can lead to significant evaluation speedup.
Note that the reason why it might work is that the supports
of f and g "approximately align with one another." But
there will be issues for large shifts. Just keep this in mind.